Removed some ancient code from uv.py:
There was a list called 'authors' in the UV-mapping function that used to collect the authors of suggestions (UvBoxes) for placing other UvBoxes on the texture, and adding their suggestions to the suggestions list. The code didn't seem to do anything. The suggestion list should be already populated with the suggestions fro all of the mapped UvBoxes so this code was just duplicating the old suggestions.